Fixed compilation issue
